move the CountVectorizer into its own file and create a stub for the VoyageAI embedding client
