+++ vectorize/count_test.go
